+++ TYPED_MOL2_EXAMPLE.mol2
@@ -1,0 +1,65 @@
+@<TRIPOS>MOLECULE
+RES
+27 27 1 0 1
+SMALL
+NO_CHARGES
+@<TRIPOS>CRYSIN
+   14.1057     9.5776     9.9538    90.0000    90.0000    90.0000 1 1
+@<TRIPOS>ATOM
+       1 C0          14.0351     9.2962    11.0772 C0            1 RES     
+       2 C0          12.8033     9.3224    10.1769 C0            1 RES     
+       3 C0          11.7631    10.3244    10.6909 C0            1 RES     
+       4 C0          10.5317    10.3530     9.7778 C0            1 RES     
+       5 C0           9.5000    11.3766    10.2763 C0            1 RES     
+       6 C0           8.2765    11.4619     9.3508 C0            1 RES     
+       7 C1           7.5013    10.1737     9.2991 C1            1 RES     
+       8 C1           6.8804     9.5986    10.4201 C1            1 RES     
+       9 C1           6.2164     8.4139    10.1362 C1            1 RES     
+      10 S2           6.4012     8.0820     8.4987 S2            1 RES     
+      11 C1           7.3225     9.4385     8.1316 C1            1 RES     
+      12 H3          14.7714     8.5664    10.6793 H3            1 RES     
+      13 H3          14.5109    10.2995    11.1092 H3            1 RES     
+      14 H3          13.7534     8.9923    12.1079 H3            1 RES     
+      15 H3          13.1184     9.6064     9.1489 H3            1 RES     
+      16 H3          12.3588     8.3036    10.1451 H3            1 RES     
+      17 H3          12.2192    11.3386    10.7229 H3            1 RES     
+      18 H3          11.4515    10.0391    11.7201 H3            1 RES     
+      19 H3          10.0743     9.3399     9.7512 H3            1 RES     
+      20 H3          10.8479    10.6297     8.7478 H3            1 RES     
+      21 H3           9.9783    12.3800    10.3054 H3            1 RES     
+      22 H3           9.1756    11.1190    11.3081 H3            1 RES     
+      23 H3           7.5984    12.2626     9.7165 H3            1 RES     
+      24 H3           8.6190    11.7457     8.3313 H3            1 RES     
+      25 H4           7.7100     9.6927     7.1541 H4            1 RES     
+      26 H5           6.9095    10.0385    11.4214 H5            1 RES     
+      27 H4           5.6658     7.8024    10.8572 H4            1 RES     
+@<TRIPOS>BOND
+       1        1        2 1
+       2        1       12 1
+       3        1       13 1
+       4        1       14 1
+       5        2        3 1
+       6        2       15 1
+       7        2       16 1
+       8        3        4 1
+       9        3       17 1
+      10        3       18 1
+      11        4        5 1
+      12        4       19 1
+      13        4       20 1
+      14        5        6 1
+      15        5       21 1
+      16        5       22 1
+      17        6        7 1
+      18        6       23 1
+      19        6       24 1
+      20        7        8 1
+      21        7       11 1
+      22        8        9 1
+      23        8       26 1
+      24        9       10 1
+      25        9       27 1
+      26       10       11 1
+      27       11       25 1
+@<TRIPOS>SUBSTRUCTURE
+       1 RES             1 RESIDUE    0 **** ROOT      0
